--- triton_dashboard/src/main/java/com/triton/msa/triton_dashboard/chat/controller/ChatController.java
@@ -1,86 +1,0 @@
-package com.triton.msa.triton_dashboard.chat.controller;
-
-import com.triton.msa.triton_dashboard.chat.entity.ChatHistory;
-import com.triton.msa.triton_dashboard.chat.service.ChatHistoryService;
-import com.triton.msa.triton_dashboard.chat.service.RagService;
-import com.triton.msa.triton_dashboard.project.entity.Project;
-import com.triton.msa.triton_dashboard.project.service.ProjectService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-import reactor.core.publisher.Flux;
-
-import java.security.Principal;
-
-@Slf4j
-@Controller
-@RequestMapping("/projects/{projectId}/chat")
-@RequiredArgsConstructor
-public class ChatController {
-
-    private final RagService ragService;
-    private final ChatHistoryService chatHistoryService;
-    private final ProjectService projectService;
-
-    @GetMapping
-    public String chatPage(@PathVariable Long projectId, Model model, Principal principal) {
-        Project project = projectService.getProject(projectId);
-
-        model.addAttribute("project", project);
-        model.addAttribute("history", chatHistoryService.getHistoryForProject(project));
-
-        return "projects/chat";
-    }
-
-//    @PostMapping("/send")
-//    public String sendMessage(@PathVariable Long projectId, @RequestParam String query, Model model, Principal principal) {
-//        Project project = projectService.getProject(projectId);
-//
-//        RagResponseDto responseDto = ragService.generateDeploymentSpec(principal.getName(), projectId, query);
-//
-//        model.addAttribute("project", project);
-//        model.addAttribute("response", responseDto);
-//        model.addAttribute("query", query);
-//
-//        model.addAttribute("history", chatHistoryService.getHistoryForProject(project));
-//
-//        return "projects/chat";
-//    }
-
-    @GetMapping(value = "/stream", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
-    @ResponseBody
-    public Flux<String> streamChatResponse(@PathVariable Long projectId, @RequestParam String query) {
-        return ragService.streamChatResponse(projectId, query);
-    }
-
-    @GetMapping("/history")
-    public String chatHistoryList(@PathVariable Long projectId, Model model) {
-        Project project = projectService.getProject(projectId);
-
-        model.addAttribute("project", project);
-        model.addAttribute("histories", chatHistoryService.getHistoryForProject(project));
-
-        return "projects/chat-history-list";
-    }
-
-    @GetMapping("/history/{historyId}")
-    public String chatHistoryDetail(@PathVariable Long projectId, @PathVariable Long historyId, Model model) {
-        Project project = projectService.getProject(projectId);
-        ChatHistory history = chatHistoryService.getHistoryById(historyId);
-
-        model.addAttribute("project", project);
-        model.addAttribute("history", history);
-
-        return "projects/chat-history-detail";
-    }
-
-    @PostMapping("/history/{historyId}/delete")
-    public String deleteHistory(@PathVariable Long historyId, @PathVariable Long projectId) {
-        chatHistoryService.deleteHistory(historyId, projectId);
-
-        return "redirect:/projects/" + projectId + "/chat/history";
-    }
-}
